Make the correct use of the protoreflect
This commit fixes the problem of protobuffer client when
running in multiple goroutines.

github.com/jhump/protoreflect has the desc.MessageDescriptor
and dynamic.Message. Before this commit, the dynamic.Message
struc was associated with the expected input. After it, the
message will be associated whit one desc.MessageDescriptor
and when some data is received, the message struct is
created. That is, each input package will have only one
dynamic.Message.