feat: read bigger messages
We have a limit of 16MB for client messages, however, we had a hard-limit for receiving 2kb chunks from the clients. This means that any message > 2kb would cause the client to be disconnected. In this commit I make it possible for the server to read chunks that are up-to MaxPacketSize in length.