refactor(agent): channel capacity as histogram
Channel capacity is a metric of each agent, that is created in each handle
so the capacity varies depending on the client/agent being used and we need
to differentiate the lows/highs of each in the metric. Adding a client label
to the Gauge is not feasible, since it'd explode cardinality. Thus, changing
the metric to be an Histogram so we can capture the highs and lows.