feat(agent): add timeout on write calls
When sending a message via agent, the write connection does not have
a default timeout, leading to blocking calls if the connections is
unhealthy or simply a bad route to it. Thus, adding a write timeout
to the writes so we fail but don't block. If deadline is exceeded then
we log an error but don't close a connection, since it might fail due
to bad route but not a broken pipe per-se. The new configuration can
be found under: `pitaya.buffer.agent.writetimeout`, with 10s by deafult.

The timeout is set using `SetWriteDeadline` function from go net pkg:
https://pkg.go.dev/net#Conn