refactor(agent): channel capacity as histogram (#450)
* refactor(agent): channel capacity as histogram (#429)

Channel capacity is a metric of each agent, that is created in each handle
so the capacity varies depending on the client/agent being used and we need
to differentiate the lows/highs of each in the metric. Adding a client label
to the Gauge is not feasible, since it'd explode cardinality. Thus, adding
a new metric as an histogram so we maintain compat until deprecating the old
gauge one

* fix(metrics): tests and usage in nats rpc (#451)

* fix(metrics): missing parenthesis on histogram
* fix(metrics): report histogram in rpc channel

* refactor(metrics): remove depreacted channel capacity gauge

* chore(go): update to 1.23