feat(session): propagate network errors on Kick
Kick might fail due to encoding and verious network errors during
write. Also, if Kick tries to write to a closed connection (fd) it
would do so without timeout, defaulting to the OS timeouts in the
order of minutes. New Kick behavior does:

- Uses writeConnection that sets write deadlines
- Propagate encode errors so client can decide to retry
- Propagate network errors so client can decide to ignore

Knowing the error is specially useful in unique_session module, which
we need to ensure existing session is always closed, regardless of a
successful Kick.