fix(agent): wrap kick error instead of metadata (#455)
Errors from generic functions used across multiple flows, like
socket write, lacks context on which flow triggered the error.
Previously, this context was added as a hardcoded reason key in
error struct, which is not ideal. Thus, changing the error to
wrap the underlying net/socket error with caller function