fix(etcd): handle revision errors in watcher
During the watchEtcd function, if we received a revision error we'd just
sleep and wait for the next event. However, these errors require the
client to establish a new connection to the server to move out of the
stale/old revision the client remained. Thus, treating these errors by
re-establishing connection