fix(nats): wait for reconnects on setup (#440)
* fix(nats): wait for reconnects on setup

If the initial connect fails, nats will spawn reconnect async
handlers. Thus, we need to wait for all reconnects to be attempted
before returning to the caller, otherwise we won't be making
use of reconnections

* fix(app): init sig chan as buffered

* fix(etcd): prevent shutdown from crashing app

If etcd module is shutdown before all connections are set up,
it will crash trying to access sd.cli where it's still nil. Thus
adding a check on shutdown

* fix CI on main

* Fix intermittent goroutine leak test failure

* remove setup of nats clusters in tests that connect to invalid addresses

* match v2 implementation

* fix remaining tests

---------

Co-authored-by: Rafael da Fonseca <rafael.fonseca@wildlifestudios.com>